requirements.txt
requests==2.31.0
python-dotenv==1.0.0
supabase==2.28.2
schedule==1.2.1
python-telegram-bot==22.7
yfinance==1.2.0
streamlit==1.44.0
plotly==5.24.0
resend
reportlab
pandas